+++ src/main/java/com/skills/qb/service/DeliveryAgentServiceImpl.java
@@ -3,10 +3,13 @@ package com.skills.qb.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.skills.qb.entity.DeliveryAgent;
 import com.skills.qb.repository.DeliveryAgentRepository;
+import com.skills.qb.response.StandardSuccessResponse;
 
 @Service
 public class DeliveryAgentServiceImpl {
@@ -18,6 +21,11 @@ public class DeliveryAgentServiceImpl {
 		return agentRepo.save(deliveryAgent);
 	}
 	
+	public DeliveryAgent deliveryAgentById(Long id) {
+//		int idl = 30/0;
+		return agentRepo.findById(id).orElse(null);
+	}
+	
 	public List<DeliveryAgent> getAllDeliveryAgents() {
 	    return agentRepo.findAll();
 	}
@@ -26,7 +34,22 @@ public class DeliveryAgentServiceImpl {
 	    return agentRepo.findByIsAvailable(isAvailable);
 	}
 
-	public void deleteById(Long id) {
-		 agentRepo.deleteById(id);
+	public String deleteById(Long id) {
+		DeliveryAgent agent = agentRepo.findById(id).orElse(null);
+		
+		 if(agent != null) {
+			 agentRepo.deleteById(id);
+		 }else {
+			 return null;
+		 }
+		 return "Successfullly deleted";
 	}
+
+	public DeliveryAgent updateAvailability(Long id, Boolean status) {
+		DeliveryAgent agent = agentRepo.findById(id).get();
+		agent.setIsAvailable(status);
+		return agentRepo.save(agent);
+	}
+
+	
 }
